\name{ref.mat}
\alias{ref.mat}
\docType{data}
\title{
  RNA-seq gene expression data set
}
\description{
  RNA-seq gene expression data set derived from 5 tissue types (lung, liver, heart, kidney, and brain) from the ArrayExpress database (E-MTAB-1733). Each tissue type is represented by 3 replicates.

}
\usage{data(ref.mat)}
\format{
  A \code{\link{matrix}} with 32431 genes and 15 samples.
}
\value{
   RNA-seq data matrix 

}


\examples{
data(ref.mat)

}
\keyword{dataset}
